sarma/backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Vector Database
pgvector==0.2.4

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1

# OpenAI & AI Services
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2  # Gemini AI

# Image Processing
pillow==10.2.0
opencv-python==4.9.0.80

# Redis Cache
redis==5.0.1
hiredis==2.3.2

# Background Tasks
celery==5.3.6
flower==2.0.1

# Monitoring & Logging
sentry-sdk==1.40.0
python-json-logger==2.0.7

# Environment & Configuration
python-dotenv==1.0.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.5.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Payment Processing
stripe==7.8.0

# Encryption
cryptography==41.0.7
